fix: align admit card identifiers to long

+++ EduOS.Core/Entities/Exams/AdmitCard.cs
@@ -5,8 +5,8 @@ namespace EduOS.Core.Entities.Exams
 {
     public class AdmitCard : BaseTenantEntity
     {
-        public int ExamId { get; set; }
-        public int StudentId { get; set; }
+        public long ExamId { get; set; }
+        public long StudentId { get; set; }
         public string AdmitNo { get; set; } = string.Empty;
         public DateTime IssueDate { get; set; }
         public bool IsIssued { get; set; } = false;
